requirements.txt
pycryptodome==3.19.0
streamlit==1.28.0
plotly==5.17.0
pandas==2.1.3
